device.c: 6: yes: MISRA.ETYPE.COMP.CAST.IMPL.WIDER.2012
Fixed R 10.7 violation by changing cast from uint16_t to uint32_t in
pm_trace call to ensure consistent essential types in bitwise OR
operation, eliminating implicit widening conversion - violation ID 2844

+++ plat/ti/k3low/common/drivers/pm/device/device.c
@@ -144,7 +144,7 @@ int32_t devices_init(void)
 						 ((((uint32_t) (-ret)) &
 						   TRACE_PM_VAL_DEV_INIT_ERR_MASK) <<
 						  TRACE_PM_VAL_DEV_INIT_ERR_SHIFT) |
-						 ((((uint16_t) idx) &
+						 ((((uint32_t) idx) &
 						   TRACE_PM_VAL_DEV_INIT_DEVICE_ID_MASK) <<
 						  TRACE_PM_VAL_DEV_INIT_DEVICE_ID_SHIFT));
 					errors = true;
